requirements.txt
pillow>=11.1.0
reportlab>=5.2.1
psutil>=7.0.0